- Renamed CISer and COSer to BinaryDeserializer and BinarySerializer - Split Connection.h/cpp into multiple files in lib/serializer directory - Cleaned up includes
