+++ client/windows/CWindowObject.h
@@ -1,9 +1,6 @@
 #pragma once
 
 #include "../gui/CIntObject.h"
-//#include "../gui/SDL_Extensions.h"
-
-//#include "../../lib/FunctionList.h"
 
 struct SDL_Surface;
 struct Rect;
