+++ lib/NetPacks.h
@@ -3,15 +3,10 @@
 #include "NetPacksBase.h"
 
 #include "BattleAction.h"
-//#include "HeroBonus.h"
 #include "mapObjects/CGHeroInstance.h"
-//#include "CCreatureSet.h"
-//#include "mapping/CMapInfo.h"
-//#include "StartInfo.h"
 #include "ConstTransitivePtr.h"
 #include "int3.h"
 #include "ResourceSet.h"
-//#include "CObstacleInstance.h"
 #include "CGameStateFwd.h"
 #include "mapping/CMapDefines.h"
 #include "CObstacleInstance.h"
