+++ lib/registerTypes/TypesMapObjects2.cpp
@@ -18,8 +18,12 @@
 #include "../NetPacks.h"
 #include "../mapObjects/CObjectClassesHandler.h"
 
+#include "../serializer/BinaryDeserializer.h"
+#include "../serializer/BinarySerializer.h"
+#include "../serializer/CTypeList.h"
 
-template void registerTypesMapObjects2<CISer>(CISer & s);
-template void registerTypesMapObjects2<COSer>(COSer & s);
+
+template void registerTypesMapObjects2<BinaryDeserializer>(BinaryDeserializer & s);
+template void registerTypesMapObjects2<BinarySerializer>(BinarySerializer & s);
 template void registerTypesMapObjects2<CTypeList>(CTypeList & s);
